backend/requirements.txt
Django
django-environ
django-cors-headers
djangorestframework
markdown
django-filter
djangorestframework-simplejwt
mysqlclient
drf-spectacular
drf_writable_nested
whitenoise
django-crum
gunicorn
